--- src/hydra/hydraClient.old-test.js
@@ -1,221 +1,0 @@
-// import {
-//   DELETE,
-//   DELETE_MANY,
-//   GET_LIST,
-//   GET_MANY,
-//   GET_MANY_REFERENCE,
-// } from 'react-admin';
-// import hydraClient, {
-//   transformJsonLdDocumentToReactAdminDocument,
-// } from './hydraClient';
-import {transformJsonLdDocumentToReactAdminDocument} from './hydraClient';
-
-describe('map a json-ld document to an admin on rest compatible document', () => {
-  const JSON_LD_DOCUMENT = {
-    '@id': '/reviews/327',
-    id: 327,
-    '@type': 'http://schema.org/Review',
-    reviewBody:
-      'Accusantium quia ipsam omnis praesentium. Neque quidem omnis perspiciatis sed. Officiis quo dolor esse nisi molestias.',
-    rating: 3,
-    itemReviewed: {
-      '@id': '/books/2',
-      id: 2,
-      '@type': 'http://schema.org/Book',
-      isbn: '9792828761393',
-      name: '000',
-      description: 'string',
-      author: 'string',
-      dateCreated: '2017-04-25T00:00:00+00:00',
-    },
-    comment: [
-      {
-        '@id': '/comments/1',
-        '@type': 'http://schema.org/Comment',
-        text: 'Lorem ipsum dolor sit amet.',
-        dateCreated: '2017-04-26T00:00:00+00:00',
-      },
-      {
-        '@id': '/comments/2',
-        '@type': 'http://schema.org/Comment',
-        text:
-          'Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
-        dateCreated: '2017-04-27T00:00:00+00:00',
-      },
-    ],
-    aNestedObject: {
-      foo: 'bar',
-    },
-  };
-
-  describe('transform the JSON-LD document in React Admin document', () => {
-    const reactAdminDocument = transformJsonLdDocumentToReactAdminDocument(
-      JSON_LD_DOCUMENT,
-    );
-
-    test('deep clone the original object', () => {
-      expect(reactAdminDocument).not.toBe(JSON_LD_DOCUMENT);
-      expect(reactAdminDocument.aNestedObject).not.toBe(
-        JSON_LD_DOCUMENT.aNestedObject,
-      );
-    });
-
-    test('add an id property equal to the original @id property', () => {
-      expect(reactAdminDocument.id).toBe(JSON_LD_DOCUMENT['@id']);
-    });
-
-    test('preserve the previous id property value in a new originId property', () => {
-      expect(reactAdminDocument.originId).toBe(JSON_LD_DOCUMENT.id);
-    });
-
-    test('an React Admin has a custom toString method', () => {
-      expect(reactAdminDocument.toString()).toBe('[object /reviews/327]');
-    });
-
-    test('transform embedded documents to their IRIs', () => {
-      expect(reactAdminDocument.itemReviewed).toBe('/books/2');
-    });
-
-    test('transform arrays of embedded documents to their IRIs', () => {
-      expect(reactAdminDocument.comment[0]).toBe('/comments/1');
-    });
-  });
-});
-
-/* 
-  Test are broken because hydraClient implementation has changed.
-  In order to find a resource, in hydraClient.js, we do :
-
-  apiSchema.resources.find(({name}) => resource === name);
-
-  => but apiSchema is undefined in our unit tests... 
-  
-  To make our tests work again, we have to change implementation of hydraClient.js
-*/
-describe('fetch data from an hydra api', () => {
-  // let MOCK_HTTP_CLIENT;
-  // beforeEach(async () => {
-  //   MOCK_HTTP_CLIENT = jest.fn();
-  // });
-
-  // test('fetch a get_list resource', async () => {
-  //   MOCK_HTTP_CLIENT.mockReturnValue(
-  //     Promise.resolve({
-  //       json: {
-  //         'hydra:member': [{'@id': 'books/5'}],
-  //       },
-  //     }),
-  //   );
-
-  //   await hydraClient('http://www.example.com', MOCK_HTTP_CLIENT)(
-  //     GET_LIST,
-  //     'books',
-  //     {
-  //       pagination: {page: 2},
-  //       sort: {field: 'id', order: 'ASC'},
-  //     },
-  //   ).then(() => {
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[0][0].href).toBe(
-  //       'http://www.example.com/books?order%5Bid%5D=ASC&page=2',
-  //     );
-  //   });
-  // });
-
-  // test('fetch a get_list resource with relative api url', async () => {
-  //   MOCK_HTTP_CLIENT.mockReturnValue(
-  //     Promise.resolve({
-  //       json: {
-  //         'hydra:member': [{'@id': 'books/5'}],
-  //       },
-  //     }),
-  //   );
-
-  //   await hydraClient('/api', MOCK_HTTP_CLIENT)(GET_LIST, 'books', {
-  //     pagination: {page: 2},
-  //     sort: {field: 'id', order: 'ASC'},
-  //   }).then(() => {
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[0][0].href).toBe(
-  //       `${window.location.origin}/api/books?order%5Bid%5D=ASC&page=2`,
-  //     );
-  //   });
-  // });
-
-  // test('fetch a get_many resource', async () => {
-  //   MOCK_HTTP_CLIENT.mockReturnValueOnce(
-  //     Promise.resolve({json: {'@id': '/books/3'}}),
-  //   ).mockReturnValue(Promise.resolve({json: {'@id': '/books/5'}}));
-
-  //   await hydraClient('http://www.example.com', MOCK_HTTP_CLIENT)(
-  //     GET_MANY,
-  //     'books',
-  //     {
-  //       ids: [3, 5],
-  //     },
-  //   ).then(response => {
-  //     expect(response.data[0].id).toEqual('/books/3');
-  //     expect(response.data[1].id).toEqual('/books/5');
-  //   });
-  // });
-
-  // test('fetch a get_many_reference resource', async () => {
-  //   MOCK_HTTP_CLIENT.mockReturnValue(
-  //     Promise.resolve({
-  //       json: {
-  //         'hydra:member': [{'@id': 'books/5'}],
-  //       },
-  //     }),
-  //   );
-
-  //   await hydraClient('/api', MOCK_HTTP_CLIENT)(GET_MANY_REFERENCE, 'books', {
-  //     target: 'author',
-  //     id: 'string',
-  //     pagination: {page: 2},
-  //     sort: {field: 'id', order: 'ASC'},
-  //     filter: {
-  //       is_published: 1,
-  //     },
-  //   }).then(response => {
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[0][0].href).toBe(
-  //       `${window.location.origin}/api/books?order%5Bid%5D=ASC&page=2&is_published=1&author=string`,
-  //     );
-  //   });
-  // });
-
-  // test('delete resource', async () => {
-  //   MOCK_HTTP_CLIENT.mockReturnValueOnce(Promise.resolve(''));
-
-  //   await hydraClient('http://www.example.com', MOCK_HTTP_CLIENT)(
-  //     DELETE,
-  //     'books',
-  //     {
-  //       id: '/books/1',
-  //     },
-  //   ).then(() => {
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[0][0].href).toBe(
-  //       'http://www.example.com/books/1',
-  //     );
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[0][1].method).toBe('DELETE');
-  //   });
-  // });
-
-  // test('delete many resources', async () => {
-  //   MOCK_HTTP_CLIENT.mockReturnValueOnce(Promise.resolve(''));
-
-  //   await hydraClient('http://www.example.com', MOCK_HTTP_CLIENT)(
-  //     DELETE_MANY,
-  //     'books',
-  //     {
-  //       ids: ['/books/1', '/books/2'],
-  //     },
-  //   ).then(() => {
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[0][0].href).toBe(
-  //       'http://www.example.com/books/1',
-  //     );
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[0][1].method).toBe('DELETE');
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[1][0].href).toBe(
-  //       'http://www.example.com/books/2',
-  //     );
-  //     expect(MOCK_HTTP_CLIENT.mock.calls[1][1].method).toBe('DELETE');
-  //   });
-  // });
-});
